requirements.txt
# Core AI & LLM
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.10
langgraph>=0.0.20
openai>=1.10.0

# Document Processing
pypdf>=3.17.0
arxiv>=2.1.0
python-docx>=1.1.0
beautifulsoup4>=4.12.0
requests>=2.31.0

# Vector Store & Embeddings
faiss-cpu>=1.7.4
chromadb>=0.4.18

# Web Interface
streamlit>=1.29.0
plotly>=5.18.0
pandas>=2.1.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
typing-extensions>=4.9.0

